release 0.0.96 source code for go

+++ services/ocr/v1/model/model_general_text_result.go
@@ -9,8 +9,8 @@ import (
 //
 type GeneralTextResult struct {
 
-	// 图片朝向，仅当detect_direction为true时，该字段有效。返回图片逆时针旋转角度，值区间为[0， 359]。当detect_direction为false时，该字段值为 -1。
-	Direction int32 `json:"direction"`
+	// 图片朝向，仅当detect_direction为true时，该字段有效。返回图片逆时针旋转角度，值区间为[0， 360)。当detect_direction为false时，该字段值为 -1。
+	Direction float32 `json:"direction"`
 
 	// 识别文字块数目。
 	WordsBlockCount int32 `json:"words_block_count"`
